pyproject.toml
[tool.poetry]
name = "smartrevbot"
version = "0.1.0"
description = ""
authors = ["skirt-owner <skirtsfield@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = {url = "https://github.com/aiogram/aiogram/archive/refs/heads/dev-3.x.zip"}
sqlalchemy = "^2.0.11"
asyncpg = "^0.27.0"
greenlet = "^2.0.2"
python-dotenv = "^1.0.0"
redis = "^4.5.4"
pydantic-settings = "^2.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
